+++ frontend/src/components/inventory-ui/EditOption.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import { ChevronLeftIcon } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { capitalize, format } from '@/lib/utils';
@@ -10,6 +10,9 @@ import ColorSelect from './ColorSelect';
 import DeletePropertyDialog from './DeletePropertyDialog';
 import { OptionType } from '@/types/options';
 import useOption from './useOptions';
+import useAssetCounter from './useAssetCounter';
+import { AssetCounterType } from '@/types/asset';
+import { useAppContext } from '@/hooks/useAppContext';
 
 interface EditOptionProps {
   option: OptionType;
@@ -19,7 +22,7 @@ interface EditOptionProps {
   className?: string;
   onCancel: () => void;
   onDelete: (optionToDelete: string) => void;
-  onUpdate: (updatedOption: OptionType) => void;
+  onUpdate: (updatedOption: OptionType, updatedAssetCounter?: AssetCounterType & { oldPrefixCode: string}) => void;
   isUpdatePending: boolean;
   onEnterPressed?: () => void;
 }
@@ -35,6 +38,12 @@ const EditOption = ({
   onEnterPressed,
 }: EditOptionProps) => {
   const [openDeleteOptionDialog, setOpenDeleteOptionDialog] = useState(false);
+  const isObject: boolean = typeof option.value === 'object';
+  const propertyIsCategory: boolean = property === 'category';
+  const [newPrefixCode, setNewPrefixCode] = propertyIsCategory ? useState('') : [undefined, () => { }];
+  const [prefixCodeError, setPrefixCodeError] = useState('');
+  const [optionError, setOptionError] = useState('');
+  const { showToast } = useAppContext();
 
   const {
     newOption: editedOption,
@@ -43,16 +52,71 @@ const EditOption = ({
     getOptionValue,
   } = useOption(option);
 
+  const {
+    getAssetCounterFromCategory,
+    updateAssetCounterInCache,
+    assetCounters
+  } = useAssetCounter(propertyIsCategory, option)
+  const assetCounter = getAssetCounterFromCategory() ?? undefined;
+  const { prefixCode: oldPrefixCode } = assetCounter ?? { prefixCode: '' };
+
   const handleCancel = () => {
     onCancel();
     setEditedOption({ property, value: '' });
   };
 
   const handleUpdate = () => {
-    onUpdate(editedOption);
+    const isPrefixCodeChanged: boolean = !!newPrefixCode && oldPrefixCode !== newPrefixCode;
+    
+    if (newPrefixCode === '') {
+      setPrefixCodeError('Prefix code can not be empty')
+      throw new Error('Prefix code can not be empty')
+    };
+
+    if (optionError) {
+      throw new Error(`${capitalize(property)} name can not be empty`)
+    }
+
+    if (assetCounters) {
+      const prefixCodeExists: boolean = !!assetCounters.find((assetCounter: AssetCounterType) => assetCounter.prefixCode === newPrefixCode);
+      if (prefixCodeExists && isPrefixCodeChanged) {
+        setPrefixCodeError('Prefix code already exists')
+        throw new Error(`Prefix code ${newPrefixCode} already exists`)
+      };
+    }
+
+    const category: string = typeof editedOption.value === 'object' ? editedOption.value.value : '';
+    
+    isPrefixCodeChanged && !!assetCounter && !!newPrefixCode ?
+      (() => {
+        onUpdate(editedOption, { ...assetCounter, category, prefixCode: newPrefixCode, oldPrefixCode })
+        updateAssetCounterInCache(assetCounter._id!, { prefixCode: newPrefixCode, category })
+      })() :
+      (() => {
+        onUpdate(editedOption)
+      })();
   };
 
   const optionToEdit = getOptionValue(editedOption.value);
+
+  useEffect(() => {
+    if (propertyIsCategory && newPrefixCode === '') {
+      setNewPrefixCode(oldPrefixCode);
+    }
+  }, [propertyIsCategory, assetCounter])
+
+  useEffect(() => {
+    if (newPrefixCode === '') {
+      setPrefixCodeError('Prefix code can not be empty')
+    } else {
+      setPrefixCodeError('')
+    }
+    if (getOptionValue(editedOption.value) === '') {
+      setOptionError(`${capitalize(property)} name can not be empty`)
+    } else {
+      setOptionError('')
+    }
+  }, [newPrefixCode, getOptionValue(editedOption.value)])
 
   return (
     <>
@@ -77,7 +141,7 @@ const EditOption = ({
         type='input'
         className='focus-visible:ring-0 focus-visible:ring-popover'
         onChange={(e) => {
-          setEditedOption({ property: property, value: e.target.value });
+          setEditedOption({ property: property, value: isObject ? {...(option.value as object), value: e.target.value} : e.target.value });
         }}
         onKeyDown={(e) => {
           if (e.key === 'Enter') {
@@ -86,6 +150,28 @@ const EditOption = ({
           }
         }}
       />
+      {optionError && <div className="text-xs text-destructive font-semibold">{optionError}</div>}
+      {propertyIsCategory && newPrefixCode !== null && (
+        <>
+          <Label>Prefix Code</Label>
+          <Input
+            value={newPrefixCode}
+            type='input'
+            className='focus-visible:ring-0 focus-visible:ring-popover'
+            onChange={(e) => {
+              const newValue = e.target.value;
+              setNewPrefixCode(newValue.toUpperCase().trim());
+            }}
+            onKeyDown={(e) => {
+              if (e.key === 'Enter') {
+                e.preventDefault();
+                onEnterPressed?.();
+              }
+            }}
+          />
+          {prefixCodeError && <div className="text-xs text-destructive font-semibold">{prefixCodeError}</div>}
+        </>
+      )}
       {colorSelect && (
         <ColorSelect
           onColorSelect={handleColorSelect}
@@ -97,9 +183,15 @@ const EditOption = ({
       <div className='flex justify-between'>
         <Button
           className='gap-2'
-          disabled={isUpdatePending}
+          disabled={!!prefixCodeError || !!optionError || isUpdatePending}
           type='button'
-          onClick={handleUpdate}
+          onClick={() => {
+            try { 
+              handleUpdate();
+            } catch (err: any) {
+              showToast({ message: err.message, type: "ERROR" })
+            }
+          }}
         >
           {isUpdatePending ? <Spinner size={18} /> : null}
           Save